Simple_bank_web _application/requirement.txt
flask==3.0.3
mysql-connector-python==9.0.0
werkzeug==3.0.3